removes knowledge about GameSetup from Game class
