+++ src/test/java/de/rabea/game/GameSetupTest.java
@@ -10,6 +10,7 @@ import org.junit.Test;
 
 import static de.rabea.game.Mark.O;
 import static de.rabea.game.Mark.X;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 public class GameSetupTest {
@@ -29,6 +30,16 @@ public class GameSetupTest {
         assertTrue(fakeUserInterface.greetUserWasCalled);
         assertTrue(gameSetup.getPlayer() instanceof HumanPlayer);
         assertTrue(gameSetup.getOpponent() instanceof HumanPlayer);
+    }
+
+    @Test
+    public void playsHuman3x3GameTwice() {
+        fakeUserInterface.provideConsoleInput("1", "3", "1", "7", "3", "4", "2", "y", "1", "3", "2", "5", "9", "7", "3", "6", "4", "8", "1", "n");
+        PlayerFactory playerFactory = new PlayerFactory(fakeUserInterface);
+        GameSetup gameSetup = new GameSetup(fakeUserInterface, playerFactory);
+        gameSetup.setUpNewGame();
+        assertTrue(fakeUserInterface.wasAskForPositionCalled());
+        assertEquals(2, fakeUserInterface.announceWinnerCalled());
     }
 
     @Test
